test: 로또 구입 금액에 대한 테스트를 LottoPurchasePrice 모델에서 수행하도록 수정
